Splits MCPage testing into two files
